Safari-609.1.11.1/LayoutTests/fast/mediastream/screencapture-disabled.html

<!DOCTYPE html>
<script src="../../resources/js-test-pre.js"></script>
<script>
    if (window.testRunner)
        window.testRunner.dumpAsText();

    if (window.internals)
        window.internals.settings.setScreenCaptureEnabled(false);

    function runTest() {
        if (!window.testRunner)
            return;

        shouldBeUndefined(navigator.mediaDevices.getDisplayMedia);
    }

    window.addEventListener("load", runTest, false);
</script>
